pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "chardet>=5.2.0",
    "docx>=0.2.4",
    "google-generativeai>=0.8.4",
    "numpy>=2.2.3",
    "openai>=1.65.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-docx>=1.1.2",
]
